requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inquirerpy==0.3.4
instagrapi==2.2.1
instaloader==4.14.2
magic-filter==1.0.12
moviepy==1.0.3
multidict==6.7.0
numpy==2.2.6
packaging==25.0
pfzy==0.3.4
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
pytest==8.3.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.22.0
yt-dlp==2026.2.4
redis==5.2.1
psycopg2-binary==2.9.10
